0.1.6

+++ spec/db_utility_spec.rb
@@ -2,35 +2,42 @@ require_relative 'spec_helper'
 
 module GemCodebreakerAmidasd
   RSpec.describe DbUtility do
-    let(:user) do
-      User.new(name: 'Amidasd')
-    end
+    let(:user) { User.new(name: 'Amidasd') }
 
-    let(:yml_db) do
-      './db/codebreaker_Test_db.yml'
-    end
+    let(:gemCodebreaker) { Game.new }
+
+    let(:yml_db) { './db/codebreaker_Test_db.yml' }
 
     before do
-      user.set_params(difficulty: :easy,
-                      total_count_attempt: 15,
-                      count_attempt: 2,
-                      total_count_hints: 2,
-                      count_hint: 0)
+      gemCodebreaker.setDifficulty(:easy)
     end
 
     after do
-      File.delete(yml_db)
+      File.delete(yml_db) if File.exist?(yml_db)
     end
+
     describe '#Test DbUtility' do
-      it 'Test add_db ' do
-        DbUtility.add_db(user, yml_db)
-        expect(File.exist?(yml_db)).equal?(true)
-        expect(DbUtility.load_yaml_db(yml_db)).equal?(true)
-        db = DbUtility.load_yaml_db(yml_db)
-        expect(db.class).equal?(Array)
-        DbUtility.add_db(user, yml_db)
-        db2 = DbUtility.load_yaml_db(yml_db)
-        expect(db.count).equal?(db2.count + 1)
+      it 'Test load ' do
+        array = DbUtility.load_yaml_db(yml_db)
+        expect(array.class).equal?(Array)
+      end
+
+      it 'Test add_in_db ' do
+        array = DbUtility.load_yaml_db(yml_db)
+        count_array_before = array.count
+        DbUtility.add_in_db(array: array, user: user, game: gemCodebreaker)
+        count_array_after = array.count
+        expect(count_array_after).equal?(count_array_before + 1)
+      end
+
+      it 'Test save ' do
+        array = DbUtility.load_yaml_db(yml_db)
+        DbUtility.add_in_db(array: array, user: user, game: gemCodebreaker)
+        DbUtility.save_yaml_db(array, yml_db)
+        array2 = DbUtility.load_yaml_db(yml_db)
+        count_array_before = array.count
+        count_array_after = array2.count
+        expect(count_array_after).equal?(count_array_before)
       end
     end
   end
